+++ Runtime/OpenXR/UserCalibration/UserCalibrationNativeTypes.cs
@@ -1,8 +1,17 @@
+// %BANNER_BEGIN%
+// ---------------------------------------------------------------------
+// %COPYRIGHT_BEGIN%
+// Copyright (c) (2024) Magic Leap, Inc. All Rights Reserved.
+// Use of this file is governed by the Software License Agreement, located here: https://www.magicleap.com/software-license-agreement-ml2
+// Terms and conditions applicable to third-party materials accompanying this distribution may also be found in the top-level NOTICE file appearing herein.
+// %COPYRIGHT_END%
+// ---------------------------------------------------------------------
+// %BANNER_END%
 ﻿using System;
 
 namespace MagicLeap.OpenXR.Features.UserCalibration
 {
-    internal enum XrUserCalibrationStructTypes : ulong
+    internal enum XrUserCalibrationStructTypes : uint
     {
         EventDataHeadsetFitChanged = 1000472000U,
         EventDataEyeCalibrationChanged = 1000472001U,
